Pick up Gwynne's ShellOut fixes, update entrypoint
